+++ app/containers/App.jsx
@@ -5,9 +5,9 @@ import { getCategories } from 'actions/categories';
 import { checkLastSession } from 'actions/session';
 import { fetchUnreadCount } from 'actions/articles';
 import { getCount } from 'helpers';
-import SettingsForm from 'containers/SettingsForm';
-import SubscribeForm from 'containers/SubscribeForm';
-import LoginForm from 'containers/LoginForm';
+import SettingsForm from 'components/SettingsForm';
+import SubscribeForm from 'components/SubscribeForm';
+import LoginForm from 'components/LoginForm';
 import Header from 'containers/Header';
 import Sidebar from 'containers/Sidebar';
 import Main from 'containers/Main';
